refactor: state를 enum class로 수정
(cherry picked from commit 69e8ba6f8314ea8910a1d46ed560c6090f1645ef)
