--- src/main/kotlin/omok/model/state/DoubleFour.kt
@@ -1,5 +1,0 @@
-package omok.model.state
-
-import omok.model.game.Board
-
-class DoubleFour(board: Board) : State(board)
