--- src/main/kotlin/omok/model/state/DoubleThree.kt
@@ -1,5 +1,0 @@
-package omok.model.state
-
-import omok.model.game.Board
-
-class DoubleThree(board: Board) : State(board)
