+++ src/main/kotlin/omok/model/state/State.kt
@@ -1,5 +1,4 @@
 package omok.model.state
-
-import omok.model.game.Board
-
-abstract class State(val board: Board)
+enum class State {
+    Win, DoubleThree, DoubleFour, Stay
+}
