--- src/main/kotlin/omok/model/state/Stay.kt
@@ -1,5 +1,0 @@
-package omok.model.state
-
-import omok.model.game.Board
-
-class Stay(board: Board) : State(board)
