--- src/main/kotlin/omok/model/state/Win.kt
@@ -1,5 +1,0 @@
-package omok.model.state
-
-import omok.model.game.Board
-
-class Win(board: Board) : State(board)
